+++ src/main/java/uk/gov/hmcts/pdda/business/services/pdda/sftp/SftpConfigHelper.java
@@ -21,9 +21,6 @@ public class SftpConfigHelper extends SftpService {
     private static final Logger LOG = LoggerFactory.getLogger(SftpConfigHelper.class);
 
     protected static final String LOG_CALLED = " called";
-    private static final String NOT_FOUND = " not found";
-    private static final String TWO_PARAMS = "{}{}";
-
     private static final String SFTP_ERROR = "SFTP Error:";
 
 
@@ -51,154 +48,6 @@ public class SftpConfigHelper extends SftpService {
     public SftpConfigHelper(EntityManager entityManager) {
         super(entityManager);
         this.entityManager = entityManager;
-    }
-
-
-    /**
-     * Set the SFTP config params.
-     * 
-     * @param sftpConfig The SFTP configuration
-     * @return The SFTP configuration
-     */
-    public SftpConfig getConfigParams(SftpConfig sftpConfig) {
-
-        if (sftpConfig.isUseKeyVault()) {
-            return getKvConfigParams(sftpConfig);
-        } else {
-            return getDbConfigParams(sftpConfig);
-        }
-    }
-
-
-    /**
-     * Set the SFTP config params from the DB.
-     * 
-     * @return The SFTP configuration
-     */
-    @SuppressWarnings("PMD.UnusedAssignment")
-    private SftpConfig getDbConfigParams(SftpConfig sftpConfig) {
-        
-        methodName = "setDBConfigParams()";
-        LOG.debug(TWO_PARAMS, methodName, LOG_CALLED);
-
-        // Firstly CP BAIS properties
-        String currentProperty = "";
-        String hostAndPort = "";
-        try {
-            currentProperty = Config.DB_CP_SFTP_USERNAME;
-            sftpConfig.setCpUsername(getMandatoryConfigValue(currentProperty));
-            LOG.debug("SFTP Cp Username: {}", sftpConfig.getCpUsername());
-
-            currentProperty = Config.DB_CP_SFTP_PASSWORD;
-            sftpConfig.setCpPassword(getMandatoryConfigValue(currentProperty));
-
-            currentProperty = Config.DB_CP_SFTP_UPLOAD_LOCATION;
-            sftpConfig
-                .setCpRemoteFolder(getMandatoryConfigValue(currentProperty));
-            LOG.debug("SFTP Cp Remote Folder: {}", sftpConfig.getCpRemoteFolder());
-
-            // Next the XHIBIT BAIS properties
-            currentProperty = Config.DB_SFTP_USERNAME;
-            sftpConfig.setXhibitUsername(getMandatoryConfigValue(currentProperty));
-            LOG.debug("SFTP XHIBIT Username: {}", sftpConfig.getCpUsername());
-
-            currentProperty = Config.DB_SFTP_PASSWORD;
-            sftpConfig.setXhibitPassword(getMandatoryConfigValue(currentProperty));
-
-            currentProperty = Config.DB_SFTP_UPLOAD_LOCATION;
-            sftpConfig
-                .setXhibitRemoteFolder(getMandatoryConfigValue(currentProperty));
-            LOG.debug("SFTP XHIBIT Remote Folder: {}", sftpConfig.getXhibitRemoteFolder());
-
-            // Now get the host and port
-            currentProperty = Config.DB_SFTP_HOST;
-            hostAndPort = getMandatoryConfigValue(currentProperty);
-            LOG.debug("SFTP Host and port: {}", hostAndPort);
-        } catch (InvalidConfigException ex) {
-            sftpConfig.setErrorMsg(currentProperty + NOT_FOUND);
-        }
-
-        // Validate the host and port
-        return validateAndSetHostAndPort(sftpConfig, hostAndPort);
-    }
-
-
-    /**
-     * Set the SFTP config params from the Key Vault.
-     * 
-     * @return The SFTP configuration
-     */
-    @SuppressWarnings("PMD.UnusedAssignment")
-    private SftpConfig getKvConfigParams(SftpConfig sftpConfig) {
-
-        methodName = "setKVConfigParams()";
-        LOG.debug(TWO_PARAMS, methodName, LOG_CALLED);
-
-        // Firstly CP BAIS properties
-        String currentProperty = "";
-        String hostAndPort = "";
-        try {
-            currentProperty = Config.KV_CP_SFTP_USERNAME;
-            sftpConfig.setCpUsername(getMandatoryEnvValue(currentProperty));
-            LOG.debug("SFTP Cp Username: {}", sftpConfig.getCpUsername());
-
-            currentProperty = Config.KV_CP_SFTP_PASSWORD;
-            sftpConfig.setCpPassword(getMandatoryEnvValue(currentProperty));
-
-            currentProperty = Config.KV_CP_SFTP_UPLOAD_LOCATION;
-            sftpConfig.setCpRemoteFolder(getMandatoryEnvValue(currentProperty));
-            LOG.debug("SFTP Cp Remote Folder: {}", sftpConfig.getCpRemoteFolder());
-
-            // Next the XHIBIT BAIS properties
-            currentProperty = Config.KV_SFTP_USERNAME;
-            sftpConfig.setCpUsername(getMandatoryEnvValue(currentProperty));
-            LOG.debug("SFTP XHIBIT Username: {}", sftpConfig.getCpUsername());
-
-            currentProperty = Config.KV_SFTP_PASSWORD;
-            sftpConfig.setCpPassword(getMandatoryEnvValue(currentProperty));
-
-            currentProperty = Config.KV_SFTP_UPLOAD_LOCATION;
-            sftpConfig.setXhibitRemoteFolder(getMandatoryEnvValue(currentProperty));
-            LOG.debug("SFTP XHIBIT Remote Folder: {}", sftpConfig.getXhibitRemoteFolder());
-
-            // Now get the host and port
-            currentProperty = Config.KV_SFTP_HOST;
-            hostAndPort = getMandatoryEnvValue(currentProperty);
-            LOG.debug("SFTP Host and port: {}", hostAndPort);
-        } catch (InvalidConfigException ex) {
-            sftpConfig.setErrorMsg(currentProperty + NOT_FOUND);
-        }
-
-        // Validate the host and port
-        return validateAndSetHostAndPort(sftpConfig, hostAndPort);
-    }
-
-
-    /**
-     * Validate and set the host and port.
-     * 
-     * @param sftpConfig The SFTP configuration
-     * @param hostAndPort The host and port
-     * @return The SFTP configuration
-     */
-    public SftpConfig validateAndSetHostAndPort(SftpConfig sftpConfig, String hostAndPort) {
-
-        LOG.debug("Validating host and port");
-        String portDelimiter = ":";
-        Integer pos = hostAndPort.indexOf(portDelimiter);
-        if (pos <= 0) {
-            sftpConfig.setErrorMsg("Host and port syntax is <Host>" + portDelimiter + "<Port>");
-        }
-        sftpConfig.setHost(hostAndPort.substring(0, pos));
-        try {
-            String strPort = hostAndPort.substring(pos + 1, hostAndPort.length());
-            sftpConfig.setPort(Integer.valueOf(strPort));
-        } catch (Exception ex) {
-            sftpConfig.setErrorMsg("Host and port contains invalid port number");
-            LOG.error("Stacktrace2:: {}", ExceptionUtils.getStackTrace(ex));
-        }
-
-        return sftpConfig;
     }
