+++ src/main/java/uk/gov/hmcts/pdda/business/services/pdda/sftp/SftpHelperUtil.java
@@ -19,6 +19,9 @@ public class SftpHelperUtil extends SftpService {
 
     private static final String USE_KEY_VAULT_PROPERTIES = "USE_KEY_VAULT_PROPERTIES";
     private static final String TEST_SFTP_DIRECTORY = "/directory/";
+    private static final String NOT_FOUND = " not found";
+    private static final String SFTP_LOG_STRING = "SFTP Host and port: {}";
+
 
     public SftpHelperUtil(EntityManager entityManager,
         XhbConfigPropRepository xhbConfigPropRepository, Environment environment,
@@ -54,7 +57,7 @@ public class SftpHelperUtil extends SftpService {
             sftpConfig.setUseKeyVault(checkWhetherToUseKeyVault());
 
             // Set the rest of the params
-            sftpConfig = getSftpConfigHelper().getConfigParams(sftpConfig);
+            sftpConfig = getConfigParams(sftpConfig);
         }
 
         return sftpConfig;
@@ -87,7 +90,7 @@ public class SftpHelperUtil extends SftpService {
      */
     private boolean checkWhetherToUseKeyVault() {
 
-        methodName = "retrieveFromBaisCp()";
+        methodName = "checkWhetherToUseKeyVault()";
         LOG.debug(methodName, LOG_CALLED);
 
         String methodName = "checkWhetherToUseKeyVault()";
@@ -102,5 +105,172 @@ public class SftpHelperUtil extends SftpService {
             LOG.error("Error fetching properties: {}", ExceptionUtils.getStackTrace(ex));
             return false;
         }
+    }
+
+
+
+    /**
+     * Set the SFTP config params.
+     * 
+     * @param sftpConfig The SFTP configuration
+     * @return The SFTP configuration
+     */
+    public SftpConfig getConfigParams(SftpConfig sftpConfig) {
+
+        if (sftpConfig.isUseKeyVault()) {
+            return getKvConfigParams(sftpConfig);
+        } else {
+            return getDbConfigParams(sftpConfig);
+        }
+    }
+
+
+    /**
+     * Set the SFTP config params from the DB.
+     * 
+     * @return The SFTP configuration
+     */
+    @SuppressWarnings("PMD.NPathComplexity")
+    private SftpConfig getDbConfigParams(SftpConfig sftpConfig) {
+
+        methodName = "getDbConfigParams()";
+        LOG.debug(TWO_PARAMS, methodName, LOG_CALLED);
+
+        // Firstly CP BAIS properties
+        try {
+            sftpConfig.setCpUsername(getMandatoryConfigValue(Config.DB_CP_SFTP_USERNAME));
+            LOG.debug("SFTP Cp Username: {}", sftpConfig.getCpUsername());
+        } catch (InvalidConfigException ex) {
+            sftpConfig.setErrorMsg(Config.DB_CP_SFTP_USERNAME + NOT_FOUND);
+        }
+        try {
+            sftpConfig
+                .setCpRemoteFolder(getMandatoryConfigValue(Config.DB_CP_SFTP_UPLOAD_LOCATION));
+            LOG.debug("SFTP Cp Remote Folder: {}", sftpConfig.getCpRemoteFolder());
+        } catch (InvalidConfigException ex) {
+            sftpConfig.setErrorMsg(Config.DB_CP_SFTP_UPLOAD_LOCATION + NOT_FOUND);
+        }
+        try {
+            sftpConfig.setXhibitPassword(getMandatoryConfigValue(Config.DB_SFTP_PASSWORD));
+        } catch (InvalidConfigException ex) {
+            sftpConfig.setErrorMsg(Config.DB_SFTP_PASSWORD + NOT_FOUND);
+        }
+
+
+
+        // Next the XHIBIT BAIS properties
+        try {
+            sftpConfig.setXhibitUsername(getMandatoryConfigValue(Config.DB_SFTP_USERNAME));
+            LOG.debug("SFTP XHIBIT Username: {}", sftpConfig.getCpUsername());
+        } catch (InvalidConfigException ex) {
+            sftpConfig.setErrorMsg(Config.DB_SFTP_USERNAME + NOT_FOUND);
+        }
+        try {
+            sftpConfig.setXhibitPassword(getMandatoryConfigValue(Config.DB_SFTP_PASSWORD));
+        } catch (InvalidConfigException ex) {
+            sftpConfig.setErrorMsg(Config.DB_SFTP_PASSWORD + NOT_FOUND);
+        }
+        try {
+            sftpConfig
+                .setXhibitRemoteFolder(getMandatoryConfigValue(Config.DB_SFTP_UPLOAD_LOCATION));
+            LOG.debug("SFTP XHIBIT Remote Folder: {}", sftpConfig.getXhibitRemoteFolder());
+        } catch (InvalidConfigException ex) {
+            sftpConfig.setErrorMsg(Config.DB_SFTP_UPLOAD_LOCATION + NOT_FOUND);
+        }
+
+        // Now get the host and port
+        String hostAndPort = "";
+        try {
+            hostAndPort = getMandatoryConfigValue(Config.DB_SFTP_HOST);
+            LOG.debug(SFTP_LOG_STRING, hostAndPort);
+        } catch (InvalidConfigException ex) {
+            sftpConfig.setErrorMsg(Config.DB_SFTP_HOST + NOT_FOUND);
+        }
+        try {
+            sftpConfig.setCpPassword(getMandatoryConfigValue(Config.DB_CP_SFTP_PASSWORD));
+        } catch (Exception ex) {
+            sftpConfig.setErrorMsg(Config.DB_CP_SFTP_PASSWORD + NOT_FOUND);
+        }
+
+        // Validate the host and port
+        return validateAndSetHostAndPort(sftpConfig, hostAndPort);
+    }
+
+    /**
+     * Validate and set the host and port.
+     * 
+     * @param sftpConfig The SFTP configuration
+     * @param hostAndPort The host and port
+     * @return The SFTP configuration
+     */
+    public SftpConfig validateAndSetHostAndPort(SftpConfig sftpConfig, String hostAndPort) {
+
+        LOG.debug("Validating host and port");
+        String portDelimiter = ":";
+        Integer pos = hostAndPort.indexOf(portDelimiter);
+        if (pos <= 0) {
+            sftpConfig.setErrorMsg("Host and port syntax is <Host>" + portDelimiter + "<Port>");
+        }
+        sftpConfig.setHost(hostAndPort.substring(0, pos));
+        try {
+            String strPort = hostAndPort.substring(pos + 1, hostAndPort.length());
+            sftpConfig.setPort(Integer.valueOf(strPort));
+        } catch (Exception ex) {
+            sftpConfig.setErrorMsg("Host and port contains invalid port number");
+            LOG.error("Stacktrace2:: {}", ExceptionUtils.getStackTrace(ex));
+        }
+
+        return sftpConfig;
+    }
+
+
+    /**
+     * Set the SFTP config params from the Key Vault.
+     * 
+     * @return The SFTP configuration
+     */
+    @SuppressWarnings("PMD.UnusedAssignment")
+    private SftpConfig getKvConfigParams(SftpConfig sftpConfig) {
+
+        methodName = "getKvConfigParams()";
+        LOG.debug(TWO_PARAMS, methodName, LOG_CALLED);
+
+        // Firstly CP BAIS properties
+        String currentProperty = "";
+        String hostAndPort = "";
+        try {
+            currentProperty = Config.KV_CP_SFTP_USERNAME;
+            sftpConfig.setCpUsername(getMandatoryEnvValue(currentProperty));
+            LOG.debug("SFTP Cp Username: {}", sftpConfig.getCpUsername());
+
+            currentProperty = Config.KV_CP_SFTP_PASSWORD;
+            sftpConfig.setCpPassword(getMandatoryEnvValue(currentProperty));
+
+            currentProperty = Config.KV_CP_SFTP_UPLOAD_LOCATION;
+            sftpConfig.setCpRemoteFolder(getMandatoryEnvValue(currentProperty));
+            LOG.debug("SFTP Cp Remote Folder: {}", sftpConfig.getCpRemoteFolder());
+
+            // Next the XHIBIT BAIS properties
+            currentProperty = Config.KV_SFTP_USERNAME;
+            sftpConfig.setCpUsername(getMandatoryEnvValue(currentProperty));
+            LOG.debug("SFTP XHIBIT Username: {}", sftpConfig.getCpUsername());
+
+            currentProperty = Config.KV_SFTP_PASSWORD;
+            sftpConfig.setCpPassword(getMandatoryEnvValue(currentProperty));
+
+            currentProperty = Config.KV_SFTP_UPLOAD_LOCATION;
+            sftpConfig.setXhibitRemoteFolder(getMandatoryEnvValue(currentProperty));
+            LOG.debug("SFTP XHIBIT Remote Folder: {}", sftpConfig.getXhibitRemoteFolder());
+
+            // Now get the host and port
+            currentProperty = Config.KV_SFTP_HOST;
+            hostAndPort = getMandatoryEnvValue(currentProperty);
+            LOG.debug(SFTP_LOG_STRING, hostAndPort);
+        } catch (InvalidConfigException ex) {
+            sftpConfig.setErrorMsg(currentProperty + NOT_FOUND);
+        }
+
+        // Validate the host and port
+        return validateAndSetHostAndPort(sftpConfig, hostAndPort);
     }
 }
